Remove QueryBuilderProperty and merge its behaviour into QueryProperty
