--- src/QueryDocuments/QueryBuilderPropertyData.ts
@@ -1,7 +1,0 @@
-import { QueryBuilderProperty } from "./QueryBuilderProperty";
-import { QueryPropertyData } from "./QueryPropertyData";
-
-
-export interface QueryBuilderPropertyData extends QueryPropertyData {
-	parent?:QueryBuilderProperty;
-}
